+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 - Upgrade Cinc Client to version to 18.4.12 from 18.2.7.
 - Upgrade Slurm to 23.11.7 (from 23.11.4).
 - Upgrade munge to version 0.5.16 (from 0.5.15).
+- Upgrade Pmix to 5.0.2 (from 4.2.9).
 - Upgrade third-party cookbook dependencies:
   - apt-7.5.22 (from apt-7.5.14)
   - openssh-2.11.12 (from openssh-2.11.3)